Make Large File Support more portable
The macro based solution has the downside of cluttering the global
namespace and requires duplication in both config headers.
Additionally ftello64 might not be available on some system where the
define check succeeds.

This solution uses a cpp file to wrap the functions adding enhanced
exists checks and additionally checking for overflow returning -1
(error) in that case as mandated by the standard for the filebuf
functions.

Finally a configure check determines general availability of LFS and if
it is not available the standard fseek/ftell functions are used.