Fix LFS detection (e.g. on Android)
(At least) the Android NDKs stdio.h does define fseeko but not if __USE_FILE_OFFSET64
is defined and the API level is below 24.
As the usage of this macro was inconsistent between the check and the
use LFS could be wrongly detected leading to a compile failure later.

Fixes #111